DBZ-1766 Unifying handling of offset commits
