+++ debezium-connector-sqlserver/src/main/java/io/debezium/connector/sqlserver/SqlServerStreamingChangeEventSource.java
@@ -343,10 +343,6 @@ public class SqlServerStreamingChangeEventSource implements StreamingChangeEvent
         return tables.toArray(new ChangeTable[tables.size()]);
     }
 
-    @Override
-    public void commitOffset(Map<String, ?> offset) {
-    }
-
     /**
      * The logical representation of a position for the change in the transaction log.
      * During each sourcing cycle it is necessary to query all change tables and then
